refactor: move BeginFiniteBurn class to new begin_finite_burn.py
